pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pygpe"
version = "1.2.0"
description = "A fast Gross-Pitaevskii equation solver for scalar, spin-1 and spin-2 BEC systems."
readme = "docs/pypi.md"
authors = [{ name = "Matt Wheeler", email = "mattwheeler15@hotmail.co.uk" }]
license = { file = "LICENSE" }
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
]
keywords = ["Gross-Pitaevskii", "Bose-Einstein", "scientific computing", "numerical methods"]
dependencies = [
    "cupy-cuda11x",
    "h5py"
]
requires-python = ">=3.10"

[project.urls]
Homepage = "https://github.com/wheelerMT/pygpe"

[tool.setuptools]
packages = ["pygpe", "pygpe.shared", "pygpe.scalar", "pygpe.spinhalf", "pygpe.spinone", "pygpe.spintwo"]

[tool.bumpver]
current_version = "1.2.0"
version_pattern = "MAJOR.MINOR.PATCH"
commit_message = "bump version {old_version} -> {new_version}"
commit = true
tag = true
push = true

[tool.bumpver.file_patterns]
"pyproject.toml" = [
    'current_version = "{version}"',
    'version = "{version}"'
]
"README.md" = [
    "{version}",
    "{pep440_version}",
]
"docs/source/conf.py" = ['release = "{version}"']

